+++ source/codegen/service_helpers.py
@@ -5,7 +5,7 @@ def get_include_guard_name(config, suffix):
     return include_guard_name.upper()
 
 def is_string_arg(parameter):
-    return parameter['type'] == 'ViChar[]' or parameter['type'] == 'ViInt8[]'
+    return parameter['type'] in ['ViChar[]', 'ViInt8[]', 'ViUInt8[]']
 
 def create_args(parameters):
     result = ''
@@ -16,7 +16,7 @@ def create_args(parameters):
       if is_output and is_string_arg(parameter):
         type_without_brackets = common_helpers.get_underlying_type_name(parameter['type'])
         result = f'{result}({type_without_brackets}*){parameter_name}.data(), '
-      elif parameter['type'] == 'ViBoolean[]':
+      elif parameter['type'] in {"ViBoolean[]", "ViSession[]"}:
         result = f'{result}{parameter_name}.data(), '
       else:
         if is_array and common_helpers.is_struct(parameter):